+++ apps/demo/app/_layout.tsx
@@ -22,6 +22,10 @@ export default function RootLayout() {
           <Stack.Screen name="index" options={{ title: 'XAUI (v1)' }} />
           <Stack.Screen name="accordion" options={{ title: 'Accordion (v1)' }} />
           <Stack.Screen name="alert" options={{ title: 'Alert (v1)' }} />
+          <Stack.Screen
+            name="autocomplete"
+            options={{ title: 'Autocomplete (v1)' }}
+          />
           <Stack.Screen name="avatar" options={{ title: 'Avatar (v1)' }} />
           <Stack.Screen name="badge" options={{ title: 'Badge (v1)' }} />
           <Stack.Screen
@@ -49,6 +53,7 @@ export default function RootLayout() {
           <Stack.Screen name="skeleton" options={{ title: 'Skeleton (v1)' }} />
           <Stack.Screen name="slider" options={{ title: 'Slider (v1)' }} />
           <Stack.Screen name="spinner" options={{ title: 'Spinner (v1)' }} />
+          <Stack.Screen name="stepper" options={{ title: 'Stepper (v1)' }} />
           <Stack.Screen name="surface" options={{ title: 'Surface (v1)' }} />
           <Stack.Screen name="switch" options={{ title: 'Switch (v1)' }} />
           <Stack.Screen name="tabs" options={{ title: 'Tabs (v1)' }} />